+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### New features
 
 - Added method `Louis.Fluency.FluentExtensions.IfNotNull`, which calls an action on an object and an additional argument only if the latter is not `null`.
+- Added overloads to `Louis.Fluency.FluentExtensions` that take simple `Action`s instead of `FluentAction`s as arguments.
 
 ### Changes to existing features